# fix(compiler): Re-enable build cache