--- a/src/compiler/cache.ts
+++ b/src/compiler/cache.ts
@@ -6,6 +6,7 @@ export class Cache implements d.Cache {
   private skip = false;
   private sys: d.CompilerSystem;
   private logger: d.Logger;
+  private cacheDir: string;
 
   constructor(private config: d.Config, private cacheFs: d.InMemoryFileSystem) {
     this.sys = config.sys;
@@ -17,16 +18,18 @@ export class Cache implements d.Cache {
       return;
     }
 
+    this.cacheDir = join(this.config.cacheDir, this.config.buildCacheDirName);
+
     if (!this.config.enableCache || !this.cacheFs) {
       this.config.logger.info(`cache optimizations disabled`);
       this.clearDiskCache();
       return;
     }
 
-    this.config.logger.debug(`cache enabled, cacheDir: ${this.config.cacheDir}`);
+    this.config.logger.debug(`cache enabled, cacheDir: ${this.cacheDir}`);
 
     try {
-      const readmeFilePath = join(this.config.cacheDir, '_README.log');
+      const readmeFilePath = join(this.cacheDir, '_README.log');
       await this.cacheFs.writeFile(readmeFilePath, CACHE_DIR_README);
     } catch (e) {
       this.logger.error(`Cache, initCacheDir: ${e}`);
@@ -121,8 +124,8 @@ export class Cache implements d.Cache {
       }
 
       const fs = this.cacheFs.sys;
-      const cachedFileNames = await fs.readDir(this.config.cacheDir);
-      const cachedFilePaths = cachedFileNames.map((f) => join(this.config.cacheDir, f));
+      const cachedFileNames = await fs.readDir(this.cacheDir);
+      const cachedFilePaths = cachedFileNames.map((f) => join(this.cacheDir, f));
 
       let totalCleared = 0;
 
@@ -148,16 +151,16 @@ export class Cache implements d.Cache {
 
   async clearDiskCache() {
     if (this.cacheFs != null) {
-      const hasAccess = await this.cacheFs.access(this.config.cacheDir);
+      const hasAccess = await this.cacheFs.access(this.cacheDir);
       if (hasAccess) {
-        await this.cacheFs.remove(this.config.cacheDir);
+        await this.cacheFs.remove(this.cacheDir);
         await this.cacheFs.commit();
       }
     }
   }
 
   private getCacheFilePath(key: string) {
-    return join(this.config.cacheDir, key) + '.log';
+    return join(this.cacheDir, key) + '.log';
   }
 
   getMemoryStats() {