@George-Payne 2023-06-02
Hi @tanner-reits ,

Great! 

 > We were wondering if you'd be able to provide the project you tested against
 
Unfortunately the project isn't open source. I'd be happy to run any benchmarks against it though, if you send instructions.