@tanner-reits 2023-06-27
Hey @George-Payne 

Just wanted to let you know that we [re-enabled build caching](#4503 ) with the release of Stencil v4! Thank you again for drawing our attention to this and doing the leg work to get it turned back on 🎉